requirements.txt
bleak
spotipy
Pillow
requests
python-dotenv
